+++ NEWS.md
@@ -10,6 +10,7 @@
 * Fixed typo if mais in `ind_managment`
 * Fixed test for winderodibility
 * Use the correct correction factor in `calc_sealing_risk` #19
+* Fix for calculating `I_P_CEC` #24
 
 ## Version 0.8.0 2019-08-02
 ### Added